+++ dbx/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.6.12"
+__version__ = "0.7.0"
