+++ dbx/callbacks.py
@@ -1,5 +1,11 @@
+from http import client as http_client
 from pathlib import Path
 from typing import Optional
+
+import typer
+from rich import print
+
+from dbx import __version__
 
 
 def verify_jinja_variables_file(_, __, value: Optional[Path]):
@@ -9,3 +15,41 @@ def verify_jinja_variables_file(_, __, value: Optional[Path]):
         if not value.exists():
             raise FileNotFoundError(f"Jinja variables file option is not empty, but file is non-existent {value}")
         return value
+
+
+def deployment_file_callback(_, value: Optional[str]) -> Path:
+    if value:
+        _candidate = Path(value)
+        if not _candidate.exists():
+            raise FileNotFoundError(f"Deployment file {_candidate} does not exist")
+    else:
+        print(":mag_right: Deployment file is not provided, searching in the [red]conf[/red] directory")
+        potential_extensions = ["json", "yml", "yaml", "json.j2", "yaml.j2", "yml.j2"]
+        _candidate = None
+        for ext in potential_extensions:
+            candidate = Path(f"conf/deployment.{ext}")
+            if candidate.exists():
+                print(f":bulb: Auto-discovery found deployment file {candidate}")
+                _candidate = candidate
+                break
+
+        raise FileNotFoundError(
+            "Auto-discovery was unable to find any deployment file in the conf directory. "
+            "Please provide file name via --deployment-file option"
+        )
+
+    print(f":ok: Deployment file {_candidate} exists and will be used for deployment")
+    return _candidate
+
+
+def version_callback(value: bool):
+    if value:
+        print(
+            f":brick:[red]Databricks[/red] e[red]X[/red]tensions aka [red]dbx[/red], version ~> [green]{__version__}[/green]"
+        )
+        raise typer.Exit()
+
+
+def debug_callback(_, value):
+    print(":bug: Debugging mode is [red]on[/red]")
+    http_client.HTTPConnection.debuglevel = 1
