+++ dbx/utils/_typer.py
@@ -1,0 +1,7 @@
+from typing import Callable
+
+from typer import Typer
+
+
+def add_callback(app: Typer, func: Callable):
+    app.callback()(func)
